feat: move RerankerModelInfo
