+++ src/models/reranking.rs
@@ -1,5 +1,7 @@
 use std::fmt::Display;
 
+use crate::RerankerModelInfo;
+
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum RerankerModel {
     /// BAAI/bge-reranker-base
@@ -44,16 +46,6 @@ pub fn reranker_model_list() -> Vec<RerankerModelInfo> {
         },
     ];
     reranker_model_list
-}
-
-/// Data struct about the available reanker models
-#[derive(Debug, Clone)]
-pub struct RerankerModelInfo {
-    pub model: RerankerModel,
-    pub description: String,
-    pub model_code: String,
-    pub model_file: String,
-    pub additional_files: Vec<String>,
 }
 
 impl Display for RerankerModel {
