+++ src/models/text_embedding.rs
@@ -1,10 +1,6 @@
-use crate::pooling::Pooling;
+use std::{collections::HashMap, fmt::Display, sync::OnceLock};
 
 use super::model_info::ModelInfo;
-
-use super::quantization::QuantizationMode;
-
-use std::{collections::HashMap, fmt::Display, sync::OnceLock};
 
 /// Lazy static list of all available models.
 static MODEL_MAP: OnceLock<HashMap<EmbeddingModel, ModelInfo<EmbeddingModel>>> = OnceLock::new();
@@ -336,78 +332,6 @@ pub fn get_model_info(model: &EmbeddingModel) -> Option<&ModelInfo<EmbeddingMode
 /// [`models_map`] instead.
 pub fn models_list() -> Vec<ModelInfo<EmbeddingModel>> {
     models_map().values().cloned().collect()
-}
-
-impl EmbeddingModel {
-    pub fn get_default_pooling_method(&self) -> Option<Pooling> {
-        match self {
-            EmbeddingModel::AllMiniLML6V2 => Some(Pooling::Mean),
-            EmbeddingModel::AllMiniLML6V2Q => Some(Pooling::Mean),
-            EmbeddingModel::AllMiniLML12V2 => Some(Pooling::Mean),
-            EmbeddingModel::AllMiniLML12V2Q => Some(Pooling::Mean),
-
-            EmbeddingModel::BGEBaseENV15 => Some(Pooling::Cls),
-            EmbeddingModel::BGEBaseENV15Q => Some(Pooling::Cls),
-            EmbeddingModel::BGELargeENV15 => Some(Pooling::Cls),
-            EmbeddingModel::BGELargeENV15Q => Some(Pooling::Cls),
-            EmbeddingModel::BGESmallENV15 => Some(Pooling::Cls),
-            EmbeddingModel::BGESmallENV15Q => Some(Pooling::Cls),
-            EmbeddingModel::BGESmallZHV15 => Some(Pooling::Cls),
-
-            EmbeddingModel::NomicEmbedTextV1 => Some(Pooling::Mean),
-            EmbeddingModel::NomicEmbedTextV15 => Some(Pooling::Mean),
-            EmbeddingModel::NomicEmbedTextV15Q => Some(Pooling::Mean),
-
-            EmbeddingModel::ParaphraseMLMiniLML12V2 => Some(Pooling::Mean),
-            EmbeddingModel::ParaphraseMLMiniLML12V2Q => Some(Pooling::Mean),
-            EmbeddingModel::ParaphraseMLMpnetBaseV2 => Some(Pooling::Mean),
-
-            EmbeddingModel::MultilingualE5Base => Some(Pooling::Mean),
-            EmbeddingModel::MultilingualE5Small => Some(Pooling::Mean),
-            EmbeddingModel::MultilingualE5Large => Some(Pooling::Mean),
-
-            EmbeddingModel::MxbaiEmbedLargeV1 => Some(Pooling::Cls),
-            EmbeddingModel::MxbaiEmbedLargeV1Q => Some(Pooling::Cls),
-
-            EmbeddingModel::GTEBaseENV15 => Some(Pooling::Cls),
-            EmbeddingModel::GTEBaseENV15Q => Some(Pooling::Cls),
-            EmbeddingModel::GTELargeENV15 => Some(Pooling::Cls),
-            EmbeddingModel::GTELargeENV15Q => Some(Pooling::Cls),
-
-            EmbeddingModel::ClipVitB32 => Some(Pooling::Mean),
-
-            EmbeddingModel::JinaEmbeddingsV2BaseCode => Some(Pooling::Mean),
-        }
-    }
-
-    /// Get the quantization mode of the model.
-    ///
-    /// Any models with a `Q` suffix in their name are quantized models.
-    ///
-    /// Currently only 6 supported models have dynamic quantization:
-    /// - Alibaba-NLP/gte-base-en-v1.5
-    /// - Alibaba-NLP/gte-large-en-v1.5
-    /// - mixedbread-ai/mxbai-embed-large-v1
-    /// - nomic-ai/nomic-embed-text-v1.5
-    /// - Xenova/all-MiniLM-L12-v2
-    /// - Xenova/all-MiniLM-L6-v2
-    ///
-    // TODO: Update this list when more models are added
-    pub fn get_quantization_mode(&self) -> QuantizationMode {
-        match self {
-            EmbeddingModel::AllMiniLML6V2Q => QuantizationMode::Dynamic,
-            EmbeddingModel::AllMiniLML12V2Q => QuantizationMode::Dynamic,
-            EmbeddingModel::BGEBaseENV15Q => QuantizationMode::Static,
-            EmbeddingModel::BGELargeENV15Q => QuantizationMode::Static,
-            EmbeddingModel::BGESmallENV15Q => QuantizationMode::Static,
-            EmbeddingModel::NomicEmbedTextV15Q => QuantizationMode::Dynamic,
-            EmbeddingModel::ParaphraseMLMiniLML12V2Q => QuantizationMode::Static,
-            EmbeddingModel::MxbaiEmbedLargeV1Q => QuantizationMode::Dynamic,
-            EmbeddingModel::GTEBaseENV15Q => QuantizationMode::Dynamic,
-            EmbeddingModel::GTELargeENV15Q => QuantizationMode::Dynamic,
-            _ => QuantizationMode::None,
-        }
-    }
 }
 
 impl Display for EmbeddingModel {
