+++ tests/embeddings.rs
@@ -7,11 +7,11 @@ use hf_hub::Repo;
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
 
 use fastembed::{
-    read_file_to_bytes, Embedding, EmbeddingModel, ImageEmbedding, ImageInitOptions, InitOptions,
-    InitOptionsUserDefined, OnnxSource, Pooling, QuantizationMode, RerankInitOptions,
-    RerankInitOptionsUserDefined, RerankerModel, SparseInitOptions, SparseTextEmbedding,
-    TextEmbedding, TextRerank, TokenizerFiles, UserDefinedEmbeddingModel,
-    UserDefinedRerankingModel, DEFAULT_CACHE_DIR,
+    read_file_to_bytes, Embedding, EmbeddingModel, ImageEmbedding, ImageEmbeddingModel,
+    ImageInitOptions, InitOptions, InitOptionsUserDefined, ModelInfo, OnnxSource, Pooling,
+    QuantizationMode, RerankInitOptions, RerankInitOptionsUserDefined, RerankerModel,
+    RerankerModelInfo, SparseInitOptions, SparseTextEmbedding, TextEmbedding, TextRerank,
+    TokenizerFiles, UserDefinedEmbeddingModel, UserDefinedRerankingModel, DEFAULT_CACHE_DIR,
 };
 
 /// A small epsilon value for floating point comparisons.
@@ -119,7 +119,7 @@ macro_rules! create_embeddings_test {
                     let embeddings = model.embed(documents.clone(), batch_size);
 
                     if matches!(
-                        (batch_size, supported_model.model.get_quantization_mode()),
+                        (batch_size, TextEmbedding::get_quantization_mode(&supported_model.model)),
                         (Some(n), QuantizationMode::Dynamic) if n < documents.len()
                     ) {
                         // For Dynamic quantization, the batch size must be greater than or equal to the number of documents
@@ -154,10 +154,12 @@ macro_rules! create_embeddings_test {
 
     };
 }
+
 create_embeddings_test!(
     name: test_batch_size_default,
     batch_size: None,
 );
+
 create_embeddings_test!(
     name: test_batch_size_less_than_document_count,
     batch_size: Some(3),
@@ -284,14 +286,11 @@ fn test_user_defined_embedding_model() {
 
 #[test]
 fn test_rerank() {
-    TextRerank::list_supported_models()
-        .par_iter()
-        .for_each(|supported_model| {
+    let test_one_model = |supported_model: &RerankerModelInfo| {
+        println!("supported_model: {:?}", supported_model);
 
-            println!("supported_model: {:?}", supported_model);
-
-        let result = TextRerank::try_new(RerankInitOptions::new(supported_model.model.clone()))
-        .unwrap();
+        let result =
+            TextRerank::try_new(RerankInitOptions::new(supported_model.model.clone())).unwrap();
 
         let documents = vec![
             "hi",
@@ -305,24 +304,35 @@ fn test_rerank() {
             .rerank("what is panda?", documents.clone(), true, None)
             .unwrap();
 
-        assert_eq!(results.len(), documents.len(), "rerank model {:?} failed", supported_model);
+        assert_eq!(
+            results.len(),
+            documents.len(),
+            "rerank model {:?} failed",
+            supported_model
+        );
 
         let option_a = "panda is an animal";
         let option_b = "The giant panda, sometimes called a panda bear or simply panda, is a bear species endemic to China.";
 
         assert!(
-            results[0].document.as_ref().unwrap() == option_a ||
-            results[0].document.as_ref().unwrap() == option_b
+            results[0].document.as_ref().unwrap() == option_a
+                || results[0].document.as_ref().unwrap() == option_b
         );
         assert!(
-            results[1].document.as_ref().unwrap() == option_a ||
-            results[1].document.as_ref().unwrap() == option_b
+            results[1].document.as_ref().unwrap() == option_a
+                || results[1].document.as_ref().unwrap() == option_b
         );
-        assert_ne!(results[0].document, results[1].document, "The top two results should be different");
+        assert_ne!(
+            results[0].document, results[1].document,
+            "The top two results should be different"
+        );
 
         // Clear the model cache to avoid running out of space on GitHub Actions.
         clean_cache(supported_model.model_code.clone())
-    });
+    };
+    TextRerank::list_supported_models()
+        .par_iter()
+        .for_each(test_one_model);
 }
 
 #[ignore]
@@ -467,26 +477,20 @@ fn test_user_defined_reranking_model() {
 
 #[test]
 fn test_image_embedding_model() {
+    let test_one_model = |supported_model: &ModelInfo<ImageEmbeddingModel>| {
+        let model: ImageEmbedding =
+            ImageEmbedding::try_new(ImageInitOptions::new(supported_model.model.clone())).unwrap();
+
+        let images = vec!["tests/assets/image_0.png", "tests/assets/image_1.png"];
+
+        // Generate embeddings with the default batch size, 256
+        let embeddings = model.embed(images.clone(), None).unwrap();
+
+        assert_eq!(embeddings.len(), images.len());
+    };
     ImageEmbedding::list_supported_models()
         .par_iter()
-        .for_each(|supported_model| {
-            let model: ImageEmbedding =
-                ImageEmbedding::try_new(ImageInitOptions::new(supported_model.model.clone()))
-                    .unwrap();
-
-            let images = vec!["tests/assets/image_0.png", "tests/assets/image_1.png"];
-
-            // Generate embeddings with the default batch size, 256
-            let embeddings = model.embed(images.clone(), None).unwrap();
-
-            assert_eq!(embeddings.len(), images.len());
-            for embedding in embeddings {
-                assert_eq!(embedding.len(), supported_model.dim);
-            }
-
-            // Clear the model cache to avoid running out of space on GitHub Actions.
-            clean_cache(supported_model.model_code.clone())
-        });
+        .for_each(test_one_model);
 }
 
 #[test]
@@ -551,17 +555,19 @@ fn clean_cache(model_code: String) {
     let cache_dir = format!("{}/{}", DEFAULT_CACHE_DIR, repo.folder_name());
     fs::remove_dir_all(cache_dir).ok();
 }
+
 // This is item "test-environment-aeghhgwpe-pro02a" of the [Aguana corpus](http://argumentation.bplaced.net/arguana/data)
 fn get_sample_text() -> String {
-    let t = "animals environment general health health general weight philosophy ethics Being vegetarian helps the environment  Becoming a vegetarian is an environmentally friendly thing to do. Modern farming is one of the main sources of pollution in our rivers. Beef farming is one of the main causes of deforestation, and as long as people continue to buy fast food in their billions, there will be a financial incentive to continue cutting down trees to make room for cattle. Because of our desire to eat fish, our rivers and seas are being emptied of fish and many species are facing extinction. Energy resources are used up much more greedily by meat farming than my farming cereals, pulses etc. Eating meat and fish not only causes cruelty to animals, it causes serious harm to the environment and to biodiversity. For example consider Meat production related pollution and deforestation  At Toronto\u{2019}s 1992 Royal Agricultural Winter Fair, Agriculture Canada displayed two contrasting statistics: \u{201c}it takes four football fields of land (about 1.6 hectares) to feed each Canadian\u{201d} and \u{201c}one apple tree produces enough fruit to make 320 pies.\u{201d} Think about it \u{2014} a couple of apple trees and a few rows of wheat on a mere fraction of a hectare could produce enough food for one person! [1]  The 2006 U.N. Food and Agriculture Organization (FAO) report concluded that worldwide livestock farming generates 18% of the planet's greenhouse gas emissions \u{2014} by comparison, all the world's cars, trains, planes and boats account for a combined 13% of greenhouse gas emissions. [2]  As a result of the above point producing meat damages the environment. The demand for meat drives deforestation. Daniel Cesar Avelino of Brazil's Federal Public Prosecution Office says \u{201c}We know that the single biggest driver of deforestation in the Amazon is cattle.\u{201d} This clearing of tropical rainforests such as the Amazon for agriculture is estimated to produce 17% of the world's greenhouse gas emissions. [3] Not only this but the production of meat takes a lot more energy than it ultimately gives us chicken meat production consumes energy in a 4:1 ratio to protein output; beef cattle production requires an energy input to protein output ratio of 54:1.  The same is true with water use due to the same phenomenon of meat being inefficient to produce in terms of the amount of grain needed to produce the same weight of meat, production requires a lot of water. Water is another scarce resource that we will soon not have enough of in various areas of the globe. Grain-fed beef production takes 100,000 liters of water for every kilogram of food. Raising broiler chickens takes 3,500 liters of water to make a kilogram of meat. In comparison, soybean production uses 2,000 liters for kilogram of food produced; rice, 1,912; wheat, 900; and potatoes, 500 liters. [4] This is while there are areas of the globe that have severe water shortages. With farming using up to 70 times more water than is used for domestic purposes: cooking and washing. A third of the population of the world is already suffering from a shortage of water. [5] Groundwater levels are falling all over the world and rivers are beginning to dry up. Already some of the biggest rivers such as China\u{2019}s Yellow river do not reach the sea. [6]  With a rising population becoming vegetarian is the only responsible way to eat.  [1] Stephen Leckie, \u{2018}How Meat-centred Eating Patterns Affect Food Security and the Environment\u{2019}, International development research center  [2] Bryan Walsh, Meat: Making Global Warming Worse, Time magazine, 10 September 2008 .  [3] David Adam, Supermarket suppliers \u{2018}helping to destroy Amazon rainforest\u{2019}, The Guardian, 21st June 2009.  [4] Roger Segelken, U.S. could feed 800 million people with grain that livestock eat, Cornell Science News, 7th August 1997.  [5] Fiona Harvey, Water scarcity affects one in three, FT.com, 21st August 2003  [6] Rupert Wingfield-Hayes, Yellow river \u{2018}drying up\u{2019}, BBC News, 29th July 2004";
+    let t = include_str!("assets/sample_text.txt");
     t.to_string()
 }
+
 #[test]
 fn test_batch_size_does_not_change_output() {
     let model = TextEmbedding::try_new(
         InitOptions::new(EmbeddingModel::AllMiniLML6V2).with_max_length(384),
     )
-    .expect("Create model succesfully");
+    .expect("Create model successfully");
 
     let sentences = vec![
         "Books are no more threatened by Kindle than stairs by elevators.",
@@ -593,7 +599,8 @@ fn test_bgesmallen1point5_match_python_counterpart() {
     let model = TextEmbedding::try_new(
         InitOptions::new(EmbeddingModel::BGESmallENV15).with_max_length(384),
     )
-    .expect("Create model succesfully");
+    .expect("Create model successfully");
+
     let text = get_sample_text();
 
     // baseline is generated in python using Xenova/bge-small-en-v1.5.onnx
@@ -631,7 +638,7 @@ fn test_allminilml6v2_match_python_counterpart() {
     let model = TextEmbedding::try_new(
         InitOptions::new(EmbeddingModel::AllMiniLML6V2).with_max_length(384),
     )
-    .expect("Create model succesfully");
+    .expect("Create model successfully");
 
     let text = get_sample_text();
